pip.requirements.txt
djangorestframework==3.4.6
Django==1.10
